refactor: dex repository
